Tests/LibWeb/Text/expected/Worker/Worker-importScripts.txt

importScripts() works as expected: YES
